consumer: rollback database transaction in case of errors
Rollback the database `Session` when an error occurs while handling a
status update message.

This fixes an issue where, if the current transaction fails for any
reason, then the consumer keeps consuming the queue without actually
updating the status of workflows and deleting the related `run-batch`
pods. This is caused by the fact that messages are acknowledged, but
querying the database fails, as the transaction was not rolled back.
This causes REANA to stop scheduling workflows, as finished workflows
are still considered as running.