feat(opensearch): capture logs from Dask cluster pods (#616)
This commit collects logs from Dask scheduler and workers and propagates
them to all REANA jobs that are using the same Dask cluster. This is not
ideal, since Dask logs can become thusly duplicated for different
workflow steps of the workflow, which could be confusing for the user.

However, when a user uses Dask to parallelise the workflow jobs, usually
the workflow steps are defined only within Dask, so this situation does
not occur. Hence we can afford doing this in usual real-life conditions.

Separating Dask scheduler and worker logs from regular Kubernetes job
logs would require a larger architectural change and is therefore
deferred to a future commit.

Closes #610